backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pytest==8.3.2
pytest-cov==5.0.0
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
